pyproject.toml
[project]
name = "odontobot-api"
version = "0.1.0"
description = "Assistente RAG WhatsApp API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-text-splitters>=0.3.0",
  "openai>=1.58.0",
  "pgvector>=0.3.0",
  "pydantic-settings>=2.6.0",
  "pypdf2>=3.0.1",
  "python-docx>=1.1.2",
  "redis>=5.2.0",
  "sqlalchemy>=2.0.0",
  "structlog>=24.4.0",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
  "factory-boy>=3.3.1",
  "mypy>=1.13.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "respx>=0.21.1",
  "ruff>=0.8.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["src/tests"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py312"
